[ BE ] (perf) 할인율 관련 모든 자료형 수정
[ BE ] (perf) 할인율 관련 모든 자료형 수정

할인율 계산할때 소수점이들어가는데, int 자료형을 쓰니 0이 반환이 되어 float 형으로 바꾸어주었습니다.

+++ 2 WEEK/SON/yebin/src/main/java/com/yebin/yebin/discount/DiscountPolicy.java
@@ -3,5 +3,5 @@ package com.yebin.yebin.discount;
 import com.yebin.yebin.member.Member;
 
 public interface DiscountPolicy {
-    int discount(Member member, int price);
+    float discount(Member member, int price);
 }
