+++ 2 WEEK/SON/yebin/src/main/java/com/yebin/yebin/discount/FixDiscountPolicy.java
@@ -7,7 +7,7 @@ public class FixDiscountPolicy implements DiscountPolicy{
     private int discountFixAmount = 1000;
 
     @Override
-    public int discount(Member member, int price) {
+    public float discount(Member member, int price) {
 
         if (member.getGrade() == Grade.VIP) {
             return discountFixAmount;
