+++ 2 WEEK/SON/yebin/src/main/java/com/yebin/yebin/discount/RateDiscountPolicy.java
@@ -4,9 +4,9 @@ import com.yebin.yebin.member.Grade;
 import com.yebin.yebin.member.Member;
 
 public class RateDiscountPolicy implements DiscountPolicy{
-    private int discountRate = 10;
+    private float discountRate = 10;
     @Override
-    public int discount(Member member, int price) {
+    public float discount(Member member, int price) {
         if (member.getGrade() == Grade.VIP){
             return price * (discountRate/100);
         }
